Replaced pit thread with callback functions

+++ Project/Sources/PIT.h
@@ -51,7 +51,7 @@ void PIT_Set(const uint32_t period, const bool restart);
  *                 FALSE if the PIT will use the new value after a trigger event.
  *  @note The function will enable the timer and interrupts for the PIT.
  */
-void PIT1_Set(const uint32_t period, const bool restart);
+void PIT1_Set(const uint64_t period, const bool restart);
 
 /*! @brief Enables or disables the PIT.
  *
